# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lykke.day",
    "description": "API for managing calendar events",
    "version": "0.1.0"
  },
  "paths": {
...
    "/days/today/context": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Get Context Today",
        "description": "Get the complete context for today.",
        "operationId": "get_context_today_days_today_context_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/days/tomorrow/context": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Get Context Tomorrow",
        "description": "Get the complete context for tomorrow.",
        "operationId": "get_context_tomorrow_days_tomorrow_context_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/days/{date}/context": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Get Context",
        "description": "Get the complete context for a specific date.",
        "operationId": "get_context_days__date__context_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/days/{date}/preview": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Preview Day",
        "description": "Preview what a day would look like if scheduled.",
        "operationId": "preview_day_days__date__preview_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/days/today/schedule": {
      "put": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Schedule Today",
        "description": "Schedule today with tasks from routines.",
        "operationId": "schedule_today_days_today_schedule_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/days/{date}/schedule": {
      "put": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Schedule Day",
        "description": "Schedule a specific day with tasks from routines.",
        "operationId": "schedule_day_days__date__schedule_put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "template_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayContextSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/days/{date}": {
      "patch": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Update Day",
        "description": "Update a day's status or template.",
        "operationId": "update_day_days__date__patch",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaySchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/days/templates": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Get Templates",
        "description": "Get all available day templates.",
        "operationId": "get_templates_days_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DayTemplateSchema"
                  },
                  "type": "array",
                  "title": "Response Get Templates Days Templates Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/days/templates/": {
      "get": {
        "tags": [
          "day",
          "scheduling",
          "planning"
        ],
        "summary": "Get Templates",
        "description": "Get all available day templates.",
        "operationId": "get_templates_days_templates__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DayTemplateSchema"
                  },
                  "type": "array",
                  "title": "Response Get Templates Days Templates  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "data": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "created_at"
        ],
        "title": "ActionSchema",
        "description": "API schema for Action value object."
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "COMPLETE",
          "DELETE",
          "EDIT",
          "NOTIFY",
          "PAUSE",
          "PUNT",
          "RESUME",
          "START",
          "VIEW"
        ],
        "title": "ActionType"
      },
      "AlarmSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "time": {
            "type": "string",
            "format": "time",
            "title": "Time"
          },
          "type": {
            "$ref": "#/components/schemas/AlarmType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "triggered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "time",
          "type"
        ],
        "title": "AlarmSchema",
        "description": "API schema for Alarm value object."
      },
      "AlarmType": {
        "type": "string",
        "enum": [
          "GENTLE",
          "FIRM",
          "LOUD",
          "SIREN"
        ],
        "title": "AlarmType"
      },
...
      "CalendarEntrySchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "calendar_id": {
            "type": "string",
            "format": "uuid",
            "title": "Calendar Id"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "calendar_id",
          "platform_id",
          "platform",
          "status",
          "starts_at",
          "frequency",
          "created_at",
          "updated_at",
          "date"
        ],
        "title": "CalendarEntrySchema",
        "description": "API schema for CalendarEntry entity."
      },
...
      "DayContextSchema": {
        "properties": {
          "day": {
            "$ref": "#/components/schemas/DaySchema"
          },
          "calendar_entries": {
            "items": {
              "$ref": "#/components/schemas/CalendarEntrySchema"
            },
            "type": "array",
            "title": "Calendar Entries"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSchema"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "DayContextSchema",
        "description": "API schema for DayContext value object."
      },
...
      "DaySchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "alarm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlarmSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/DayTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DayTemplateSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "date",
          "status"
        ],
        "title": "DaySchema",
        "description": "API schema for Day entity."
      },
      "DayStatus": {
        "type": "string",
        "enum": [
          "UNSCHEDULED",
          "SCHEDULED",
          "IN_PROGRESS",
          "COMPLETE"
        ],
        "title": "DayStatus"
      },
      "DayTag": {
        "type": "string",
        "enum": [
          "WEEKEND",
          "VACATION",
          "WORKDAY"
        ],
        "title": "DayTag"
      },
...
      "DayTemplateSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "alarm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlarmSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "routine_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Routine Ids"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "user_id"
        ],
        "title": "DayTemplateSchema",
        "description": "API schema for DayTemplate entity."
      },
...
      "DayUpdateSchema": {
        "properties": {
          "alarm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlarmSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DayStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DayTag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          }
        },
        "type": "object",
        "title": "DayUpdateSchema",
        "description": "API schema for Day update requests."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCategory": {
        "type": "string",
        "enum": [
          "HYGIENE",
          "GROOMING",
          "SLEEP",
          "HEALTH",
          "FITNESS",
          "MEDICAL",
          "NUTRITION",
          "COOKING",
          "HOUSE",
          "CLEANING",
          "MAINTENANCE",
          "ORGANIZATION",
          "WORK",
          "MEETING",
          "PROFESSIONAL",
          "FAMILY",
          "SOCIAL",
          "RELATIONSHIP",
          "SHOPPING",
          "ERRAND",
          "COMMUTE",
          "TRAVEL",
          "ENTERTAINMENT",
          "HOBBY",
          "RECREATION",
          "EDUCATION",
          "LEARNING",
          "FINANCE",
          "BILLS",
          "TECHNOLOGY",
          "DIGITAL",
          "PET",
          "PLANNING",
          "ADMIN"
        ],
        "title": "TaskCategory"
      },
...
      "TaskDefinitionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "type": {
            "$ref": "#/components/schemas/TaskType"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "type",
          "user_id"
        ],
        "title": "TaskDefinitionSchema",
        "description": "API schema for TaskDefinition entity."
      },
...
      "TaskFrequency": {
        "type": "string",
        "enum": [
          "DAILY",
          "CUSTOM_WEEKLY",
          "WEEKLY",
          "ONCE",
          "YEARLY",
          "MONTHLY",
          "BI_WEEKLY",
          "WORK_DAYS",
          "WEEKENDS"
        ],
        "title": "TaskFrequency"
      },
...
      "TaskScheduleSchema": {
        "properties": {
          "available_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Time"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "timing_type": {
            "$ref": "#/components/schemas/TimingType"
          }
        },
        "type": "object",
        "required": [
          "timing_type"
        ],
        "title": "TaskScheduleSchema",
        "description": "API schema for TaskSchedule value object."
      },
      "TaskSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_definition": {
            "$ref": "#/components/schemas/TaskDefinitionSchema"
          },
          "category": {
            "$ref": "#/components/schemas/TaskCategory"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskScheduleSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "routine_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Id"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TaskTag"
            },
            "type": "array",
            "title": "Tags"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "scheduled_date",
          "name",
          "status",
          "task_definition",
          "category",
          "frequency"
        ],
        "title": "TaskSchema",
        "description": "API schema for Task entity."
      },
      "TaskStatus": {
        "type": "string",
        "enum": [
          "COMPLETE",
          "NOT_READY",
          "READY",
          "PUNT",
          "NOT_STARTED",
          "PENDING"
        ],
        "title": "TaskStatus"
      },
      "TaskTag": {
        "type": "string",
        "enum": [
          "AVOIDANT",
          "FORGETTABLE",
          "IMPORTANT",
          "URGENT",
          "FUN"
        ],
        "title": "TaskTag"
      },
      "TaskType": {
        "type": "string",
        "enum": [
          "MEAL",
          "WORK",
          "MEETING",
          "EXERCISE",
          "EVENT",
          "SOCIAL",
          "CHORE",
          "ERRAND",
          "SHOPPING",
          "PERSONAL_CARE",
          "ACTIVITY",
          "ENTERTAINMENT",
          "LEARNING",
          "COMMUTE",
          "TRAVEL",
          "APPOINTMENT",
          "COMMUNICATION",
          "FINANCIAL",
          "MAINTENANCE",
          "PLANNING",
          "TECHNOLOGY"
        ],
        "title": "TaskType"
      },
      "TimingType": {
        "type": "string",
        "enum": [
          "DEADLINE",
          "FIXED_TIME",
          "TIME_WINDOW",
          "FLEXIBLE"
        ],
        "title": "TimingType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lykke_auth"
      }
    }
  }
}